+++ src/main/resources/asset/~act/js/jquery.ext.js
@@ -131,9 +131,9 @@ jQuery.each(["getJSON", "postJSON", "putJSON", "deleteJSON", "patchJSON"], funct
         if (method.startsWith("get")) {
             var hasParam = url.contains("?");
             if (!hasParam) {
-                url = url +"?now="+ new Date().getTime();
+                url = url +"?_now="+ new Date().getTime();
             } else {
-                url = url +"&now="+ new Date().getTime();
+                url = url +"&_now="+ new Date().getTime();
             }
         }
         var setup = {